consume biovil-t weights
