+++ hi-ml-multimodal/test_multimodal/vlp/test_zero_shot_classification.py
@@ -8,7 +8,7 @@ from torchvision.datasets.utils import check_integrity
 
 from health_multimodal.image import ImageInferenceEngine
 from health_multimodal.image.data.transforms import create_chest_xray_transform_for_inference
-from health_multimodal.image.model.model import get_biovil_resnet
+from health_multimodal.image.model.pretrained import get_biovil_resnet
 from health_multimodal.text.utils import get_cxr_bert_inference
 from health_multimodal.vlp.inference_engine import ImageTextInferenceEngine
